Admin and guest areas

+++ routes/web.php
@@ -6,6 +6,8 @@ use App\Http\Controllers\Posts;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\CarsController;
 use App\Http\Controllers\BrandsController;
+use App\Http\Controllers\Guest\CarsGuestController;
+use Illuminate\Routing\RouteRegistrar;
 
 Route::prefix('/auth')->group(function(){
     Route::controller(SessionsController::class)->group(function(){
@@ -15,25 +17,32 @@ Route::prefix('/auth')->group(function(){
         Route::post('/register', 'store')->name('auth.register.store');
     });
 });
+
+
 Route::post('/logout', [LogoutController::class, 'store'])->name('logout'); 
 
-Route::middleware('auth')->group(function(){
-Route::get('/cars', [CarsController::class, 'index'])->name('cars.showAll'); //вывод на гоавную всех машин
-Route::get('/cars/test', [CarsController::class, 'test'])->name('cars.test'); //Тест
-Route::get('/cars/check', [CarsController::class, 'check']); //Тест
-Route::get('/cars/create', [CarsController::class, 'create'])->name('cars.create'); //Переход на страничку создания
-Route::post('/cars', [CarsController::class, 'store'])->name('cars.store'); //Создание новой сущности (просто записи) в базе
-Route::post('/cars/{car}/comment', [CarsController::class, 'addComment'])->name('cars.addComment'); //Добавление комментария к машине 
-Route::get('/cars/trash', [CarsController::class, 'showTrashCars'])->name('cars.showTrashCars'); //Показ удаленных машин
-Route::put('/cars/{car}/restore', [CarsController::class, 'restore'])->name('cars.restore'); //Восстановление 1 машины
-Route::patch('/cars/{car}', [CarsController::class, 'update'])->name('cars.update'); //Редактирование 1 записи
-Route::get('/cars/{car}/redaction', [CarsController::class, 'redactionById'])->name('cars.redactionById'); //Дубль над чем работаем 
-Route::get('/cars/{car}', [CarsController::class, 'show'])->name('cars.showById');
-Route::delete('/cars/{id}/destroyForever', [CarsController::class, 'destroyForever'])->name('cars.destroyForever'); //Окончательное удаление 1 машины
-Route::delete('/cars/{car}', [CarsController::class, 'destroy'])->name('cars.delete');
+
+Route::prefix('admin')->middleware(['auth', 'isAdmin'])->group(function(){
+Route::get('/cars', [CarsController::class, 'index'])->name('admin.cars.showAll'); //вывод на главную всех машин
+Route::get('/cars/create', [CarsController::class, 'create'])->name('admin.cars.create'); //Переход на страничку создания
+Route::post('/cars', [CarsController::class, 'store'])->name('admin.cars.store'); //Создание новой сущности (просто записи) в базе
+Route::post('/cars/{car}/comment', [CarsController::class, 'addComment'])->name('admin.cars.addComment'); //Добавление комментария к машине 
+Route::get('/cars/trash', [CarsController::class, 'showTrashCars'])->name('admin.cars.showTrashCars'); //Показ удаленных машин
+Route::put('/cars/{car}/restore', [CarsController::class, 'restore'])->name('admin.cars.restore'); //Восстановление 1 машины
+Route::patch('/cars/{car}', [CarsController::class, 'update'])->name('admin.cars.update'); //Редактирование 1 записи
+Route::get('/cars/{car}/redaction', [CarsController::class, 'admin.redactionById'])->name('admin.cars.redactionById'); //Дубль над чем работаем 
+Route::get('/cars/{car}', [CarsController::class, 'show'])->name('admin.cars.showById');
+Route::delete('/cars/{id}/destroyForever', [CarsController::class, 'destroyForever'])->name('admin.cars.destroyForever'); //Окончательное удаление 1 машины
+Route::delete('/cars/{car}', [CarsController::class, 'destroy'])->name('admin.cars.delete');
 });
 
-
+Route::prefix('cars')->middleware('auth')->group(function(){
+    Route::controller(CarsGuestController::class)->group(function(){
+        Route::get('/', 'index')->name('cars.showAll');
+        Route::get('/{car}', 'show')->name('cars.showById');
+        Route::post('/{car}/comment', 'addComment')->name('cars.addComment');
+    });
+});
 
 
 Route::get('/brands/{brand}/description',[BrandsController::class, 'brandDescription'])->name('brands.brandDescription');
